resolving all conflict and succesfully run appliaction

+++ LMS.Application/Services/Loan/EmiCalculatorService.cs
@@ -70,7 +70,7 @@ namespace LMS.Application.Services.Loan
                 if (month == tenureMonths)
                 {
                     principalPart = remainingPrincipal;
-                    emi = (double)principalPart + interestPart;
+                    emi = (decimal)(principalPart + interestPart);
                 }
                 
                 remainingPrincipal -= principalPart;
